+++ src/components/mapbox/Map.js
@@ -1,6 +1,6 @@
 import { useRef, useState } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
-import MapGL, { Layer, Source } from 'react-map-gl';
+import MapGL from 'react-map-gl';
 import 'mapbox-gl/dist/mapbox-gl.css';
 
 import {
@@ -9,6 +9,7 @@ import {
   initialViewState,
   ZoomRange,
 } from '@config/map-config';
+import useGetRoutes from '@hooks/useGetRoutes';
 import { getReverseAddress } from '@services/photo';
 import { setCoordinateFile } from '@store/custom';
 import '@styles/mapbox/Map.scss';
@@ -17,6 +18,7 @@ import '@styles/mapbox/Marker.scss';
 import CustomPopup from './CustomPopup';
 import TripMarkers from './TripMarkers';
 import WriteMarkers from './WriteMarkers';
+import CustomRoutes from './CustomRoutes';
 
 const Map = () => {
   const MapGLRef = useRef();
@@ -25,6 +27,7 @@ const Map = () => {
 
   const { isCustomMode } = useSelector((state) => state.custom);
   const { markers, type } = useSelector((state) => state.marker);
+  const { routes } = useGetRoutes(markers);
   const dispatch = useDispatch();
 
   const onShowPopup = async (e) => {
@@ -43,23 +46,6 @@ const Map = () => {
     setShowPopup(false);
   };
 
-  const routesMarker = markers
-    .filter((marker) => marker.status !== 'unspecified')
-    .map((marker) => {
-      const { latitude, longitude } = marker.coordinate;
-      return [longitude, latitude];
-    });
-
-  // 경로 표시
-  const routes = {
-    type: 'Feature',
-    properties: {},
-    geometry: {
-      type: 'LineString',
-      coordinates: [...routesMarker],
-    },
-  };
-
   return (
     <div className='map-container'>
       <MapGL
@@ -72,24 +58,11 @@ const Map = () => {
         {...ZoomRange}
       >
         {type === 'TRIP' && <TripMarkers markers={markers} />}
-        {type === 'WRITE' && <WriteMarkers markers={markers} />}
         {type === 'WRITE' && (
-          <Source id='polylineLayer' type='geojson' data={routes}>
-            <Layer
-              id='lineLayer'
-              type='line'
-              source='my-data'
-              layout={{
-                'line-join': 'round',
-                'line-cap': 'round',
-              }}
-              paint={{
-                'line-color': '#0095f6',
-                'line-width': 4,
-                'line-dasharray': [0, 3, 3],
-              }}
-            />
-          </Source>
+          <>
+            <WriteMarkers markers={markers} />
+            <CustomRoutes routes={routes} />
+          </>
         )}
         {showPopup && (
           <CustomPopup
